+++ src/hooks/playerlife/Oxygen.cpp
@@ -25,7 +25,7 @@ class $modify(OxygenPlayLayer, PlayLayer) {
         float oxygenLevel = 50.f;
 
         ProgressBar* oxygenBar = nullptr;
-        CCLabelBMFont* oxygenLabel = nullptr;
+        Label* oxygenLabel = nullptr;
     };
 
     HORRIBLE_SETUP_INTERFACE_FUNC {
@@ -59,7 +59,7 @@ class $modify(OxygenPlayLayer, PlayLayer) {
 
         auto const o2 = fmt::format("o2\n{}%", static_cast<uint8_t>(f->oxygenLevel));
         if (!f->oxygenLabel) {
-            f->oxygenLabel = CCLabelBMFont::create(o2.c_str(), font::big);
+            f->oxygenLabel = Label::create(o2.c_str(), font::big);
             f->oxygenLabel->setColor(colors::cyan);
             f->oxygenLabel->setAnchorPoint({0.f, 1.f});
             f->oxygenLabel->setPosition({2.f, (getScaledContentHeight() / 2.f) - (f->oxygenBar->getScaledContentWidth() / 2.f) - 1.25f});